requirements.txt
django>=5.2.9
djangorestframework>=3.16.1
google-generativeai>=0.3.0
python-dotenv>=1.0.0
